Pipeline/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
altair=4.2.0=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pyhd8ed1ab_0
audioread=2.1.9=pypi_0
backcall=0.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.10.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blinker=1.4=pypi_0
bpemb=0.3.3=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py38h578d9bd_2
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
conllu=4.4.1=pypi_0
cryptography=36.0.0=py38h9ce1e76_0
ctc-segmentation=1.7.1=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
cython=0.29.28=pypi_0
dacite=1.6.0=pypi_0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.0.0=pypi_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=pypi_0
dill=0.3.4=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
fastjsonschema=2.15.3=pypi_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.6.0=pyhd8ed1ab_0
flair=0.10=pypi_0
fonttools=4.31.2=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=pypi_0
fsspec=2022.3.0=pypi_0
ftfy=6.1.1=pypi_0
gdown=3.12.2=pypi_0
gensim=4.1.2=pypi_0
gettext=0.19.8.1=h0b5b191_1005
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
huggingface_hub=0.4.0=pyhd8ed1ab_0
hypothesis=6.41.0=pyhd8ed1ab_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=3.10.1=pypi_0
importlib-resources=5.6.0=pypi_0
importlib_metadata=4.11.3=hd8ed1ab_1
install=1.3.5=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.12.1=pypi_0
ipython=8.2.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
janome=0.4.2=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.1=pypi_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h7f8727e_0
jsonschema=4.4.0=pypi_0
jupyter-client=7.2.1=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
kiwisolver=1.4.2=pypi_0
konoha=4.6.5=pypi_0
lame=3.100=h7b6447c_0
langdetect=1.0.9=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libflac=1.3.4=h27087fc_0
libgcc-ng=11.2.0=h1d223b6_14
libgomp=11.2.0=h1d223b6_14
libidn2=2.3.2=h7f8727e_0
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
librosa=0.9.1=pypi_0
libsndfile=1.0.31=h9c3ff4c_1
libstdcxx-ng=11.2.0=he4da1e4_14
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
llvmlite=0.38.0=pypi_0
lxml=4.8.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
more-itertools=8.8.0=pypi_0
mpld3=0.3=pypi_0
multidict=6.0.2=pypi_0
multiprocess=0.70.12.2=pypi_0
nbclient=0.5.13=pypi_0
nbconvert=6.4.5=pypi_0
nbformat=5.3.0=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=pypi_0
nettle=3.7.3=hbbd107a_1
notebook=6.4.10=pypi_0
numba=0.55.1=pypi_0
numpy=1.21.5=pypi_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.2.0=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h166bdaf_0
overrides=3.1.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=pypi_0
pandocfilters=1.5.0=pypi_0
parallelformers=1.2.4=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
pooch=1.6.0=pypi_0
prometheus-client=0.13.1=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.20.0=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=7.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyctcdecode=0.3.0=pyhd8ed1ab_0
pydeck=0.7.1=pypi_0
pygments=2.11.2=pypi_0
pygtrie=2.4.2=pyhd3deb0d_0
pympler=1.0.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pyhd8ed1ab_0
pypi-kenlm=0.1.20210121=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
python-levenshtein=0.12.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.12.0.dev20220404=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0=py38h0a891b7_4
pyzmq=22.3.0=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.3.15=py38h0a891b7_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
resampy=0.2.2=pypi_0
resemblyzer=0.1.1.dev0=pypi_0
responses=0.18.0=pypi_0
rsa=4.8=pypi_0
sacremoses=0.0.49=pyhd8ed1ab_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
segtok=1.5.11=pypi_0
semver=2.13.0=pypi_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.95=pypi_0
sentry-sdk=1.5.8=pypi_0
seqeval=1.2.2=pypi_0
setproctitle=1.2.2=pypi_0
setuptools=58.0.4=py38h06a4308_0
shortuuid=1.0.8=pypi_0
simpletransformers=0.63.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
smart-open=5.2.1=pypi_0
smmap=5.0.0=pypi_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
soupsieve=2.3.1=pypi_0
sqlite=3.38.2=hc218d9a_0
sqlitedict=2.0.0=pypi_0
stack-data=0.2.0=pypi_0
streamlit=1.8.1=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.3=pypi_0
testpath=0.6.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=py38hb317417_1
toml=0.10.2=pypi_0
toolz=0.11.2=pypi_0
torchaudio=0.12.0.dev20220404=py38_cu102
torchvision=0.13.0.dev20220404=py38_cu102
tornado=6.1=pypi_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.1=pypi_0
transformers=4.16.2=pyhd8ed1ab_0
typing=3.7.4.3=pypi_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022.1=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.8=pyhd3eb1b0_0
validators=0.18.2=pypi_0
wandb=0.12.11=pypi_0
watchdog=2.1.7=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
webrtcvad=2.0.10=pypi_0
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=pypi_0
wikipedia-api=0.5.4=pypi_0
wrapt=1.14.0=pypi_0
x264=1!157.20191217=h7b6447c_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7f98852_2
yarl=1.7.2=pypi_0
yaspin=2.1.0=pypi_0
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
